backend/requirements.txt
flask
flask-socketio
flask-cors
eventlet
python-dotenv
gunicorn
